[Viewer] Replace multi args "," delimiter by "|" to allow comma in path. Raise exception is

+++ python/jiminy_py/src/jiminy_py/meshcat/recorder.py
@@ -177,7 +177,7 @@ def meshcat_recorder(meshcat_url, request_shm, message_shm):
                 while request_shm.value != "quit":  # [>Python3.8] while (request := request_shm.value) != "quit":
                     request = request_shm.value
                     if request != "":
-                        args = map(str.strip, message_shm.value.split(","))
+                        args = map(str.strip, message_shm.value.split("|"))
                         if request == "take_snapshot":
                             width, height = map(int, args)
                             coro = capture_frame_async(client, width, height)
@@ -300,13 +300,13 @@ class MeshcatRecorder:
 
     def capture_frame(self, width=None, height=None):
         self._send_request("take_snapshot",
-            message=f"{width if width is not None else -1},"\
+            message=f"{width if width is not None else -1}|"\
                     f"{height if height is not None else -1}")
         return self.__shm['message'].value
 
     def start_video_recording(self, fps, width, height):
         self._send_request("start_record",
-            message=f"{fps},{width},{height}", timeout=10.0)
+            message=f"{fps}|{width}|{height}", timeout=10.0)
         self.is_recording = True
 
     def add_video_frame(self):
@@ -335,6 +335,9 @@ class MeshcatRecorder:
             raise RuntimeError(
                 "No video being recorded at the moment. "\
                 "Please start recording and add frames before saving.")
+        if "|" in path:
+            raise ValueError(
+                "'|' character is not supported in video export path.")
         path = os.path.abspath(pathlib.Path(path).with_suffix('.webm'))
         if os.path.exists(path):
             os.remove(path)
